# backend/docs/sabi-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "sabi-api",
    "version": "1.0.0",
    "description": "Sistema de Gestión de Activos y Usuarios de la Empresa",
    "contact": {
      "name": "API Support",
      "email": "support@empresa.com"
    }
  },
  "servers": [
    {
      "url": "https://sabi.bispace.site/api",
      "description": "Servidor principal de sabi-api"
    }
  ],
  "paths": {
    "/usr": {
      "get": {
        "summary": "Gestión de usuarios del sistema",
        "operationId": "get_usr",
        "tags": [
          "usr"
        ],
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "description": "Filtrar por estado (activo/inactivo)",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/activo": {
      "get": {
        "summary": "Lista de activos de la empresa",
        "operationId": "get_activo",
        "tags": [
          "activo"
        ],
        "parameters": [
          {
            "name": "categoria",
            "in": "query",
            "description": "Filtrar por categoría de activo",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/cantidad": {
      "get": {
        "summary": "Consultas de cantidad y estadísticas",
        "operationId": "get_cantidad",
        "tags": [
          "cantidad"
        ],
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "description": "Tipo de consulta (usuarios, activos, etc.)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}